requirements.txt
#requests>=2,<3
#msal>=0.6.1,<2
requests
msal
azure.identity
pandas
sshtunnel
python-dotenv
email.mime
smtplib